refactor: split _doctor.py (1245 lines) into 8 focused sub-modules
Extracts all 30 _check_* functions from cli/_doctor.py into domain sub-modules
(_doctor_types, _doctor_mcp, _doctor_hooks, _doctor_install, _doctor_config,
_doctor_runtime, _doctor_env, _doctor_features, _doctor_fleet). _doctor.py
becomes a facade (221 lines, ≤250 limit). Adds CD5 line-limit guard and 18
structural importability tests.
