+++ tests/cli/test_doctor_split.py
@@ -10,6 +10,7 @@ import pytest
 pytestmark = [pytest.mark.layer("cli"), pytest.mark.small]
 
 _CLI_TESTS = Path(__file__).parent
+_CLI_SRC = Path(__file__).parent.parent.parent / "src" / "autoskillit" / "cli"
 
 
 def _has_pytestmark_cli(path: Path) -> bool:
@@ -73,3 +74,75 @@ def test_doctor_core_does_not_contain_quota_cache_class():
     assert "TestCheckQuotaCacheSchema" not in class_names, (
         "TestCheckQuotaCacheSchema must be moved to test_doctor_migration.py"
     )
+
+
+def test_doctor_types_file_exists():
+    assert (_CLI_SRC / "_doctor_types.py").exists()
+
+
+def test_doctor_mcp_file_exists():
+    assert (_CLI_SRC / "_doctor_mcp.py").exists()
+
+
+def test_doctor_hooks_file_exists():
+    assert (_CLI_SRC / "_doctor_hooks.py").exists()
+
+
+def test_doctor_install_file_exists():
+    assert (_CLI_SRC / "_doctor_install.py").exists()
+
+
+def test_doctor_config_file_exists():
+    assert (_CLI_SRC / "_doctor_config.py").exists()
+
+
+def test_doctor_runtime_file_exists():
+    assert (_CLI_SRC / "_doctor_runtime.py").exists()
+
+
+def test_doctor_env_file_exists():
+    assert (_CLI_SRC / "_doctor_env.py").exists()
+
+
+def test_doctor_features_file_exists():
+    assert (_CLI_SRC / "_doctor_features.py").exists()
+
+
+def test_doctor_fleet_file_exists():
+    assert (_CLI_SRC / "_doctor_fleet.py").exists()
+
+
+def test_doctor_facade_exports_run_doctor():
+    from autoskillit.cli._doctor import run_doctor  # noqa: F401
+
+
+def test_doctor_submodule_mcp_importable():
+    from autoskillit.cli._doctor_mcp import _check_mcp_server_registered  # noqa: F401
+
+
+def test_doctor_submodule_hooks_importable():
+    from autoskillit.cli._doctor_hooks import _check_hook_registry_drift  # noqa: F401
+
+
+def test_doctor_submodule_install_importable():
+    from autoskillit.cli._doctor_install import _check_stale_entry_points  # noqa: F401
+
+
+def test_doctor_submodule_config_importable():
+    from autoskillit.cli._doctor_config import _check_config_layers_for_secrets  # noqa: F401
+
+
+def test_doctor_submodule_runtime_importable():
+    from autoskillit.cli._doctor_runtime import _check_quota_cache_schema  # noqa: F401
+
+
+def test_doctor_submodule_env_importable():
+    from autoskillit.cli._doctor_env import _check_ambient_session_type_leaf  # noqa: F401
+
+
+def test_doctor_submodule_features_importable():
+    from autoskillit.cli._doctor_features import _check_feature_dependencies  # noqa: F401
+
+
+def test_doctor_submodule_fleet_importable():
+    from autoskillit.cli._doctor_fleet import _check_stale_fleet_state  # noqa: F401
